requirements.txt
numpy>=1.19.5
pandas>=1.1.5
matplotlib>=3.3.4
opencv-python>=4.5.1
pillow>=8.1.0
tensorflow>=2.5.0
scikit-learn>=0.24.1
imagehash>=4.2.0
seaborn>=0.11.1
tqdm>=4.61.0